refactor for pylint
